- Add reminder options to the new command - In item lists also show if a item has work log items - Add an --overview option to ls that shows all outstanding tasks in all projects - Allow users to close items as part of the shell description - Add a --ifopen option to remind that will only remind if the item is still open - Add --done option to the remind command that will also close an item after reminding

+++ lib/gwtf/commands/remind_command.rb
@@ -19,39 +19,52 @@ command [:remind, :rem] do |c|
   c.default_value false
   c.switch [:send]
 
+  c.desc 'Mark item done after sending reminder'
+  c.default_value false
+  c.switch [:done]
+
+  c.desc 'Only alert if the item is still marked open'
+  c.default_value false
+  c.switch [:ifopen]
+
   c.action do |global_options,options,args|
     raise "Please supply an item ID to remind about" if args.empty?
 
     unless options[:send]
       raise "Please specify a valid at() time specification" unless args.size > 2
 
-      at = args[1..-1].join(" ")
-
       STDOUT.sync
 
       print "Creating reminder at job for item #{args.first}: "
-      system "echo gwtf --project='%s' remind --recipient='%s' --send %s|at %s" % [ global_options[:project], options[:recipient], args.first, at]
+
+      @items.load_item(args.first).schedule_reminer(args[1..-1].join(" "), options[:recipient], options[:done], options[:ifopen])
     else
       item = @items.load_item(args.first)
 
-      begin
-        tmp = Tempfile.new("gwtf")
-        tmp.write(item.summary)
-        tmp.rewind
+      unless options[:ifopen] && item.closed?
+        begin
+          tmp = Tempfile.new("gwtf")
+          tmp.write(item.summary)
+          tmp.rewind
 
-        if global_options[:project] == "default"
-          subject = "Reminder for item %s" % [ args.first ]
-        else
-          subject = "Reminder for item %s in %s project" % [ args.first, global_options[:project] ]
+          if global_options[:project] == "default"
+            subject = "Reminder for item %s" % [ args.first ]
+          else
+            subject = "Reminder for item %s in %s project" % [ args.first, global_options[:project] ]
+          end
+
+          system("cat #{tmp.path} | mail -s '#{subject}' '#{options[:recipient]}'")
+
+          if options[:done]
+            item.record_work("Closing item as part of scheduled reminder")
+            item.close
+            item.save
+          end
+        ensure
+          tmp.close
+          tmp.unlink
         end
-
-        system("cat #{tmp.path}|mail -s '#{subject}' '#{options[:recipient]}'")
-      ensure
-        tmp.close
-        tmp.unlink
       end
     end
   end
 end
-
-
